Add triple pane view along with modal for upload (#22)
* Linting of files and creating panes

* Add Multipane to project

* Display a triple pane with project contents

This commit implements a triple pane which shows Original Language,
Target Language, and Gateway Language. At the current moment,
only Target Language is shown because that is what's inputed from
the Translation Studio import. In the future, Original Language
will be brought in, along with Gateway Language

* Add a modal for the drag and drop upload feature

This commits adds a component that holds the upload drag and drop.
It also adds all the needed event handlers, and actions to the
appropriate functions.

* Fixed suggested fixes and linted all files